+++ src/Dracodeum/Kit/Interfaces/Stringable.php
@@ -7,8 +7,6 @@
 
 namespace Dracodeum\Kit\Interfaces;
 
-use Dracodeum\Kit\Options\Text as TextOptions;
-
 /** This interface defines a method to cast an object to a string. */
 interface Stringable
 {
@@ -16,10 +14,10 @@ interface Stringable
 	/**
 	 * Cast this object to a string.
 	 * 
-	 * @param \Dracodeum\Kit\Options\Text|null $text_options [default = null]
-	 * <p>The text options instance to use.</p>
+	 * @param coercible:options(Dracodeum\Kit\Options\Text)|null $text_options [default = null]
+	 * <p>The text options to use.</p>
 	 * @return string
 	 * <p>This object cast to a string.</p>
 	 */
-	public function toString(?TextOptions $text_options = null): string;
+	public function toString($text_options = null): string;
 }